@@ -34,16 +34,18 @@
   "homepage": "https://github.com/liquid-labs/format-and-lint#readme",
   "dependencies": {
     "@babel/eslint-parser": "^7.22.15",
+    "@eslint/compat": "^1.1.1",
     "@eslint/js": "^8.56.0",
     "@liquid-labs/sdlc-resource-babel-and-rollup": "^1.0.0-alpha.9",
-    "@stylistic/eslint-plugin": "^2.6.4",
-    "eslint": "^8.50.0",
+    "@stylistic/eslint-plugin": "^2.8.0",
+    "eslint": "^9.10.0",
     "eslint-config-standard": "^17.1.0",
     "eslint-plugin-import": "^2.29.1",
     "eslint-plugin-jsdoc": "^48.0.4",
     "eslint-plugin-n": "^16.6.2",
     "eslint-plugin-node": "^11.1.0",
     "eslint-plugin-promise": "^6.1.1",
+    "find-plus": "^1.1.1",
     "globals": "^14.0.0",
     "jest": "^29.7.0",
     "prettier": "^3.3.3"
@@ -62,5 +64,8 @@
         "src/test/data/**"
       ]
     }
+  },
+  "peerDependencies": {
+    "command-line-args": "^6.0.0"
   }
 }
